Cleanup resistor-color exercise files for new structure requirements.
