+++ exercises/practice/resistor-color/.docs/instructions.md
