+++ exercises/practice/resistor-color/ResistorColorTest.php
@@ -1,0 +1,64 @@
+<?php
+
+/*
+ * By adding type hints and enabling strict type checking, code can become
+ * easier to read, self-documenting and reduce the number of potential bugs.
+ * By default, type declarations are non-strict, which means they will attempt
+ * to change the original type to match the type specified by the
+ * type-declaration.
+ *
+ * In other words, if you pass a string to a function requiring a float,
+ * it will attempt to convert the string value to a float.
+ *
+ * To enable strict mode, a single declare directive must be placed at the top
+ * of the file.
+ * This means that the strictness of typing is configured on a per-file basis.
+ * This directive not only affects the type declarations of parameters, but also
+ * a function's return type.
+ *
+ * For more info review the Concept on strict type checking in the PHP track
+ * <link>.
+ *
+ * To disable strict typing, comment out the directive below.
+ */
+
+declare(strict_types=1);
+
+class ResistorColorTest extends PHPUnit\Framework\TestCase
+{
+    public static function setUpBeforeClass(): void
+    {
+        require_once 'ResistorColor.php';
+    }
+
+    public function testBlackColorCode(): void
+    {
+        $this->assertEquals(colorCode("black"), 0);
+    }
+
+    public function testOrangeColorCode(): void
+    {
+        $this->assertEquals(colorCode("orange"), 3);
+    }
+
+    public function testWhiteColorCode(): void
+    {
+        $this->assertEquals(colorCode("white"), 9);
+    }
+
+    public function testColors(): void
+    {
+        $this->assertEquals(COLORS, [
+            "black",
+            "brown",
+            "red",
+            "orange",
+            "yellow",
+            "green",
+            "blue",
+            "violet",
+            "grey",
+            "white"
+        ]);
+    }
+}
