--- exercises/resistor-color/example.php
@@ -1,8 +1,0 @@
-<?php
-
-const COLORS = ["black", "brown", "red", "orange", "yellow", "green", "blue", "violet", "grey", "white"];
-
-function colorCode($color)
-{
-    return array_search($color, COLORS);
-}
