--- exercises/resistor-color/resistor-color.php
@@ -1,1 +1,0 @@
-<?php
